[TD-225] fix compiler warning.

+++ src/query/src/qUtil.c
@@ -369,7 +369,7 @@ int32_t getNumOfTotalRes(SGroupResInfo* pGroupResInfo) {
     return 0;
   }
 
-  return taosArrayGetSize(pGroupResInfo->pRows);
+  return (int32_t) taosArrayGetSize(pGroupResInfo->pRows);
 }
 
 static int64_t getNumOfResultWindowRes(SQueryRuntimeEnv* pRuntimeEnv, SResultRow *pResultRow) {
